+++ ask/commands.py
@@ -1,7 +1,7 @@
 import sublime
 import sublime_plugin
 import textwrap
-from pieces_os_client import (
+from .._pieces_lib.pieces_os_client import (
 	QGPTTaskPipelineForCodeFix,
 	QGPTTaskPipeline,
 	QGPTTaskPipelineForCodeCommentation,
@@ -75,6 +75,8 @@ class PiecesAskQuestionCommand(sublime_plugin.TextCommand):
 
 	def on_done(self,description):
 		self.description = description
+		if not self.description:
+			self.description = "No description provided"
 		sublime.set_timeout_async(self.on_done_async,0)
 
 	def on_done_async(self):