+++ event_listener.py
@@ -4,7 +4,6 @@ import sublime_plugin
 from .assets.markdown_handler import PiecesHandleMarkdownCommand
 from .assets.list_assets import PiecesListAssetsCommand
 from .settings import PiecesSettings
-from .base_websocket import BaseWebsocket
 from .copilot.ask_command import copilot